backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
pydantic==2.10.3
python-dotenv==1.0.1
httpx==0.28.1
transformers==4.47.0
torch==2.5.1
readability-lxml==0.8.1
beautifulsoup4==4.12.3
pytest==8.3.4
pytest-asyncio==0.24.0
pydantic-settings==2.6.1
